Revert "feat: add API to publish community template (#27134)"
This reverts commit b203d9481cff785f9a3f4730fe8dde72aa92e1f5.
