+++ app/server/appsmith-server/src/main/java/com/appsmith/server/configurations/CloudServicesConfig.java
@@ -17,9 +17,6 @@ public class CloudServicesConfig {
     @Value("${appsmith.cloud_services.password}")
     private String password;
 
-    @Value("${appsmith.cloud_services.template_upload_auth_header}")
-    private String templateUploadAuthHeader;
-
     @Autowired
     public void setBaseUrl(@Value("${appsmith.cloud_services.base_url:}") String value) {
         baseUrl = StringUtils.isEmpty(value) ? "https://cs.appsmith.com" : value;
