+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/ApplicationTemplate.java
@@ -1,7 +1,6 @@
 package com.appsmith.server.dtos;
 
 import com.appsmith.external.models.BaseDomain;
-import jakarta.validation.constraints.Email;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -35,11 +34,4 @@ public class ApplicationTemplate extends BaseDomain {
     private Boolean featured;
     private List<String> tags;
     private Boolean allowPageImport;
-    // This flag denotes whether a template is an official template
-    // or a community template
-    private Boolean isCommunityTemplate;
-    // This will point to the email of the template's author. This cannot be
-    // null if the template is a community template
-    @Email
-    private String authorEmail;
 }
