--- app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/CommunityTemplateDTO.java
@@ -1,20 +1,0 @@
-package com.appsmith.server.dtos;
-
-import com.appsmith.external.models.BaseDomain;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class CommunityTemplateDTO extends BaseDomain {
-    String applicationId;
-    String workspaceId;
-    String branchName;
-    String title;
-    String headline;
-    String description;
-    List<String> useCases;
-    String authorEmail;
-}
