--- app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/CommunityTemplateUploadDTO.java
@@ -1,11 +1,0 @@
-package com.appsmith.server.dtos;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CommunityTemplateUploadDTO {
-    ApplicationTemplate applicationTemplate;
-    ApplicationJson appJson;
-}
