+++ app/server/appsmith-server/src/main/resources/application.properties
@@ -80,7 +80,6 @@ appsmith.cloud_services.base_url = ${APPSMITH_CLOUD_SERVICES_BASE_URL:}
 appsmith.cloud_services.signature_base_url = ${APPSMITH_CLOUD_SERVICES_SIGNATURE_BASE_URL:}
 appsmith.cloud_services.username = ${APPSMITH_CLOUD_SERVICES_USERNAME:}
 appsmith.cloud_services.password = ${APPSMITH_CLOUD_SERVICES_PASSWORD:}
-appsmith.cloud_services.template_upload_auth_header = ${APPSMITH_CLOUD_SERVICES_TEMPLATE_UPLOAD_AUTH}
 github_repo = ${APPSMITH_GITHUB_REPO:}
 
 # MANDATORY!! No default properties are being provided for encryption password and salt for security.
